+++ src/shapes/ActiveSelection.ts
@@ -174,14 +174,6 @@ export class ActiveSelection extends Group {
   }
 
   /**
-   * Check if this group or its parent group are caching, recursively up
-   * @return {Boolean}
-   */
-  isOnACache() {
-    return false;
-  }
-
-  /**
    * Renders controls and borders for the object
    * @param {CanvasRenderingContext2D} ctx Context to render on
    * @param {Object} [styleOverride] properties to override the object style