+++ test/visual/clippath.js
@@ -30,37 +30,37 @@
     percentage: 0.06,
   });
 
-  function clipping01(canvas, callback) {
-    var clipPath = new fabric.Circle({ radius: 50, strokeWidth: 40, top: -50, left: -50, fill: 'transparent' });
-    var obj = new fabric.Rect({ top: 0, left: 0, strokeWidth: 0, width: 200, height: 200, fill: 'rgba(0,255,0,0.5)'});
-    obj.clipPath = clipPath;
-    canvas.add(obj);
-    canvas.renderAll();
-    callback(canvas.lowerCanvasEl);
-  }
+  // function clipping01(canvas, callback) {
+  //   var clipPath = new fabric.Circle({ radius: 50, strokeWidth: 40, top: -50, left: -50, fill: 'transparent' });
+  //   var obj = new fabric.Rect({ top: 0, left: 0, strokeWidth: 0, width: 200, height: 200, fill: 'rgba(0,255,0,0.5)'});
+  //   obj.clipPath = clipPath;
+  //   canvas.add(obj);
+  //   canvas.renderAll();
+  //   callback(canvas.lowerCanvasEl);
+  // }
 
-  tests.push({
-    test: 'A clippath ignores fill and stroke for drawing, not positioning',
-    code: clipping01,
-    golden: 'clipping01.png',
-    percentage: 0.06,
-  });
+  // tests.push({
+  //   test: 'A clippath ignores fill and stroke for drawing, not positioning',
+  //   code: clipping01,
+  //   golden: 'clipping01.png',
+  //   percentage: 0.06,
+  // });
 
-  function clipping02(canvas, callback) {
-    var clipPath = new fabric.Circle({ radius: 50, strokeWidth: 40, top: -50, left: -50, fill: '' });
-    var obj = new fabric.Rect({ top: 0, left: 0, strokeWidth: 0, width: 200, height: 200, fill: 'rgba(0,255,0,0.5)'});
-    obj.clipPath = clipPath;
-    canvas.add(obj);
-    canvas.renderAll();
-    callback(canvas.lowerCanvasEl);
-  }
+  // function clipping02(canvas, callback) {
+  //   var clipPath = new fabric.Circle({ radius: 50, strokeWidth: 40, top: -50, left: -50, fill: '' });
+  //   var obj = new fabric.Rect({ top: 0, left: 0, strokeWidth: 0, width: 200, height: 200, fill: 'rgba(0,255,0,0.5)'});
+  //   obj.clipPath = clipPath;
+  //   canvas.add(obj);
+  //   canvas.renderAll();
+  //   callback(canvas.lowerCanvasEl);
+  // }
 
-  tests.push({
-    test: 'falsy values for fill are handled',
-    code: clipping02,
-    golden: 'clipping01.png',
-    percentage: 0.06,
-  });
+  // tests.push({
+  //   test: 'falsy values for fill are handled',
+  //   code: clipping02,
+  //   golden: 'clipping01.png',
+  //   percentage: 0.06,
+  // });
 
   function clipping1(canvas, callback) {
     var zoom = 20;